Binary Tree with Templates
